fix compiler warnings in utils/ on OSX

+++ utils/bpsplit/bpappend.c
@@ -15,6 +15,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <stdint.h>
+#include <inttypes.h>
 
 #ifndef __USE_LARGEFILE64
 #define __USE_LARGEFILE64
@@ -213,12 +214,12 @@ int read_indexes(char *filename, bool input) {
         printf (DIVIDER);
         printf ("%s file %s:\n", (input ? "Input" : "Output"), filename);
         printf ("Process Groups Index:\n");
-        printf ("Process Groups Index Offset = %llu\n", bp->pg_index_offset);
-        printf ("Process Groups Index Size   = %llu\n", bp->pg_size);
-        printf ("Variable Index Offset       = %llu\n", bp->vars_index_offset);
-        printf ("Variable Index Size         = %llu\n", bp->vars_size);
-        printf ("Attribute Index Offset      = %llu\n", bp->attrs_index_offset);
-        printf ("Attribute Index Size        = %llu\n", bp->attrs_size);
+        printf ("Process Groups Index Offset = %" PRIu64 "\n", bp->pg_index_offset);
+        printf ("Process Groups Index Size   = %" PRIu64 "\n", bp->pg_size);
+        printf ("Variable Index Offset       = %" PRIu64 "\n", bp->vars_index_offset);
+        printf ("Variable Index Size         = %" PRIu64 "\n", bp->vars_size);
+        printf ("Attribute Index Offset      = %" PRIu64 "\n", bp->attrs_index_offset);
+        printf ("Attribute Index Size        = %" PRIu64 "\n", bp->attrs_size);
     }
 
     // set global pointers 
@@ -273,8 +274,8 @@ int copy_file( const char *filein, const char *fileout) {
     while (bytes_read > 0) {
         bytes_written = write( outf, (void *) buf, bytes_read);
         if (bytes_written != bytes_read) {
-            fprintf(stderr, "Error: could not write %d bytes to output file %s at offset 0: %s\n", 
-                    bytes_read, fileout, strerror(errno));
+            fprintf(stderr, "Error: could not write %" PRId64 " bytes to output file %s at offset 0: %s\n",
+                    (long long)bytes_read, fileout, strerror(errno));
             close(inf);
             close(outf);
             return 4;
@@ -283,13 +284,13 @@ int copy_file( const char *filein, const char *fileout) {
         bytes_read = read( inf, (void *)buf, COPYBUFSIZE);
     }
     if (bytes_read < 0) {
-        fprintf(stderr, "Error at reading input file %s from offset %llu: %s\n", 
+        fprintf(stderr, "Error at reading input file %s from offset %" PRIu64 ": %s\n",
                 filein, bytes_copied, strerror(errno));
         close(inf);
         close(outf);
         return 2;
     }
-    if (verbose>1) printf("  copied %llu (0x%llx) bytes of data into %s\n", bytes_copied, bytes_copied, fileout);
+    if (verbose>1) printf("  copied %" PRIu64 " (0x%llx) bytes of data into %s\n", bytes_copied, bytes_copied, fileout);
 
     close(inf);
     close(outf);
@@ -394,14 +395,14 @@ int recover(int f) {
     // write old index into a buffer
     adios_write_index_v1 (&buffer, &buffer_size, &buffer_offset, index_start, 
                           idx);
-    if (verbose>1) fprintf(stderr, "  original index size %llu 0x%llx\n", buffer_offset, buffer_offset);
+    if (verbose>1) fprintf(stderr, "  original index size %" PRIu64 " 0x%llx\n", buffer_offset, buffer_offset);
     adios_write_version_v1 (&buffer, &buffer_size, &buffer_offset);
 
     adios_clear_index_v1 (idx);
     adios_free_index_v1 (idx);
 
     // write index buffer out
-    if (verbose>1) fprintf(stderr, "  write %llu 0x%llx bytes of indexes...\n", buffer_offset, buffer_offset);
+    if (verbose>1) fprintf(stderr, "  write %" PRIu64 " 0x%llx bytes of indexes...\n", buffer_offset, buffer_offset);
     bytes_written = write (f, buffer, buffer_offset);
     free(buffer);
     if (bytes_written == -1) {
@@ -409,7 +410,7 @@ int recover(int f) {
         return 2;
     }
     if (bytes_written != buffer_offset) {
-        fprintf(stderr, "  Error: could not write complete original index into output file: %zu bytes from %llu\n"
+        fprintf(stderr, "  Error: could not write complete original index into output file: %zu bytes from %" PRIu64 "\n"
                 "Index in output file will be damaged.\n",
                 bytes_written, buffer_offset);
         return 2;
@@ -451,26 +452,26 @@ int append_in_to_out( const char *fileout, const char *filein) {
     off_t    pos;
     lseek64 (in_bp->f, 0, SEEK_SET);
     if (verbose>1) 
-        printf("  seek in output to end of group data, %llu (0x%llx) \n", 
+        printf("  seek in output to end of group data, %" PRIu64 " (0x%llx) \n",
                 out_bp->pg_index_offset, out_bp->pg_index_offset);
     pos = lseek64 (f, out_bp->pg_index_offset, SEEK_SET); // seek behind groups, overwrite old index
     if (pos != out_bp->pg_index_offset) {
-            fprintf(stderr, "Error: cannot seek to offset %llu (0x%llx) in file %s: %s\n", 
+            fprintf(stderr, "Error: cannot seek to offset %" PRIu64 " (0x%llx) in file %s: %s\n",
                     out_bp->pg_index_offset, out_bp->pg_index_offset, filein, strerror(errno));
             close(f);
             return 2;
     }
-    if (verbose>1) printf("  copy data from input file, %llu bytes\n", in_bp->pg_index_offset);
+    if (verbose>1) printf("  copy data from input file, %" PRIu64 " bytes\n", in_bp->pg_index_offset);
     while (bytes_copied < bytes_to_copy) {
         bytes_read = read( in_bp->f, (void *)buf, COPYBUFSIZE);
         if (bytes_read < 0) {
-            fprintf(stderr, "Error at reading input file %s from offset %llu: %s\n", 
+            fprintf(stderr, "Error at reading input file %s from offset %" PRIu64 ": %s\n",
                     filein, bytes_copied, strerror(errno));
             recover(f);
             close(f);
             return 3;
         } else if (bytes_read == 0) {
-            fprintf(stderr, "Error unexpected end of input file %s at offset %llu: %s\n", 
+            fprintf(stderr, "Error unexpected end of input file %s at offset %" PRIu64 ": %s\n",
                     filein, bytes_copied, strerror(errno));
             recover(f);
             close(f);
@@ -482,8 +483,8 @@ int append_in_to_out( const char *fileout, const char *filein) {
 
         bytes_written = write( f, (void *) buf, bytes_read);
         if (bytes_written != bytes_read) {
-            fprintf(stderr, "Error: could not write %d bytes to output file %s at offset %llu: %s\n", 
-                    bytes_read, fileout, out_bp->pg_index_offset+bytes_copied, strerror(errno));
+            fprintf(stderr, "Error: could not write %" PRId64 " bytes to output file %s at offset %" PRIu64 ": %s\n",
+                    (long long)bytes_read, fileout, out_bp->pg_index_offset+bytes_copied, strerror(errno));
             recover(f);
             close(f);
             return 4;
@@ -491,7 +492,7 @@ int append_in_to_out( const char *fileout, const char *filein) {
 
         bytes_copied += bytes_written;
     }
-    if (verbose>1) printf("  written %llu (0x%llx) bytes of data into %s\n", bytes_copied, bytes_copied, fileout);
+    if (verbose>1) printf("  written %" PRIu64 " (0x%llx) bytes of data into %s\n", bytes_copied, bytes_copied, fileout);
 
     // Append input indexes into the output indexes
     char * buffer = 0;
@@ -504,18 +505,18 @@ int append_in_to_out( const char *fileout, const char *filein) {
     idx->vars_root = out_vars_root;
     idx->attrs_root = out_attrs_root;
 
-    if (verbose>1) printf("  index starts at %llu (0x%llx)\n", index_start, index_start);
+    if (verbose>1) printf("  index starts at %" PRIu64 " (0x%llx)\n", index_start, index_start);
 
     // merge in old indicies
     adios_merge_index_v1 (idx,
                           in_pg_root, in_vars_root, in_attrs_root, 0);
     adios_write_index_v1 (&buffer, &buffer_size, &buffer_offset, index_start, 
                           idx);
-    if (verbose>1) printf("  index size %llu 0x%llx\n", buffer_offset, buffer_offset);
+    if (verbose>1) printf("  index size %" PRIu64 " 0x%llx\n", buffer_offset, buffer_offset);
     adios_write_version_v1 (&buffer, &buffer_size, &buffer_offset);
 
     // write index buffer out
-    if (verbose>1) printf("  write %llu 0x%llx bytes of indexes into %s\n", buffer_offset, buffer_offset, fileout);
+    if (verbose>1) printf("  write %" PRIu64 " 0x%llx bytes of indexes into %s\n", buffer_offset, buffer_offset, fileout);
     bytes_written = write (f, buffer, buffer_offset);
 
     if (verbose>1) printf("  written %zu 0x%zx bytes of indexes into %s\n", bytes_written, bytes_written, fileout);
